requirements.txt
langchain
langchain-openai
openai>=1.10.0
python-dotenv